backend/requirements.txt
django
djangorestframework
pandas
matplotlib
reportlab
django-cors-headers
